requirements.txt
# Core ML libraries
torch>=1.12.0
torchvision>=0.13.0
timm>=0.6.12

# Data processing
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.3.0

# Visualization and analysis
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter and notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Model export and inference
onnx>=1.12.0
onnxruntime>=1.12.0

# Web application
gradio>=3.0.0

# Utilities
tqdm>=4.64.0
scikit-learn>=1.1.0
